Allow peak freezing etc. without showing toolbar
The toolbar is unnecessary for many actions when
using keyboard shortcuts. When other tools are
active the window can get a little cluttered.

Perhaps this should be the default? In this
implementation it requires an explicit call of
FXMLController.initPeakSlider(false)

In this implementation calling
FXMLController.initPeakSlider(false) will hide an
existing toolbar. It is easy to re-add, but this
may be undesired behavior nonetheless.