Fixed bug in `nmrfx-analyst-gui/src/main/scripts/filter.py`
Part of this script is meant to read the file
`target/myFilter.properties`, and extract the classpath.

The file's structure is of the form
`classpath={path-1}:{path-2}:{path-3}:...`.

The previous form of `filter.py` neglected the 1st path in the list
(splitting was done at every `:`); this edit fixes that.